+++ src/main/java/com/example/goodsprice/activity/application/domain/service/ActivityLogService.java
@@ -2,17 +2,13 @@ package com.example.goodsprice.activity.application.domain.service;
 
 import static com.example.goodsprice.activity.infrastructure.config.ActivityLogConstants.ENTITY_NAME;
 
-import com.example.goodsprice.activity.application.domain.model.ActivityLogAction;
 import com.example.goodsprice.activity.application.domain.model.ActivityLogDomain;
-import com.example.goodsprice.activity.application.domain.model.ActivityLogType;
 import com.example.goodsprice.activity.application.port.in.ActivityLogInPort;
+import com.example.goodsprice.activity.application.port.in.dto.ActivityLogCriteria;
 import com.example.goodsprice.activity.application.port.out.ActivityLogRepositoryPort;
 import com.example.goodsprice.common.constant.ErrorCodes;
-import com.example.goodsprice.common.dto.PageRequestDto;
 import com.example.goodsprice.common.dto.PageResponse;
 import com.example.goodsprice.common.service.AbstractGenericService;
-import java.time.OffsetDateTime;
-import java.util.Objects;
 import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -43,18 +39,7 @@ public class ActivityLogService extends AbstractGenericService<ActivityLogDomain
   }
 
   @Override
-  public PageResponse<ActivityLogDomain> findAll(
-      int page,
-      int size,
-      String sortBy,
-      String sortDirection,
-      ActivityLogType type,
-      ActivityLogAction action,
-      OffsetDateTime startDate,
-      OffsetDateTime endDate) {
-    var pageRequest = new PageRequestDto(page, size, sortBy, sortDirection);
-    var start = Objects.nonNull(startDate) ? startDate.toLocalDateTime() : null;
-    var end = Objects.nonNull(endDate) ? endDate.toLocalDateTime() : null;
-    return activityLogRepository.findAll(pageRequest, type, action, start, end);
+  public PageResponse<ActivityLogDomain> findAll(ActivityLogCriteria criteria) {
+    return activityLogRepository.findAll(criteria);
   }
 }